security-dashboard/requirements.txt
# Security Monitoring Dashboard Requirements
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2

# Database and caching
asyncpg==0.29.0
aioredis==2.0.1

# HTTP client and data processing
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6

# Configuration and data formats
pyyaml==6.0.1
python-dotenv==1.0.0

# AWS SDK (for real AWS integration)
boto3==1.34.0
botocore==1.34.0

# Azure SDK (for real Azure integration)
azure-identity==1.15.0
azure-mgmt-security==6.0.0
azure-mgmt-monitor==6.0.0
azure-storage-blob==12.19.0

# GCP SDK (for real GCP integration)
google-cloud-security-center==1.26.0
google-cloud-monitoring==2.17.0
google-cloud-logging==3.8.0

# Security and encryption
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Additional utilities
click==8.1.7
rich==13.7.0
schedule==1.2.0